less abstraction: remove StoreProxy
